+++ pages/operations.php
@@ -616,6 +616,54 @@ if($_GET['action'] == 'importCompounds'){
 
 }
 
+//IMPORT CATEGORIES
+if($_GET['action'] == 'importCategories'){
+	if (!file_exists($tmp_path)) {
+		mkdir($tmp_path, 0777, true);
+	}
+	
+	if (!is_writable($tmp_path)) {
+		$result['error'] = "Upload directory not writable. Make sure you have write permissions.";
+		echo json_encode($result);
+		return;
+	}
+	
+	$target_path = $tmp_path.basename($_FILES['jsonFile']['name']); 
+
+	if(move_uploaded_file($_FILES['jsonFile']['tmp_name'], $target_path)) {
+    	$data = json_decode(file_get_contents($target_path), true);
+		if(!$data['ingCategory']){
+			$result['error'] = "JSON File seems invalid. Please make sure you importing the right file";
+			echo json_encode($result);
+			return;
+		}
+		
+		foreach ($data['ingCategory'] as $d ){				
+			
+			$s = mysqli_query($conn, "INSERT INTO `ingCategory` (`name`,`notes`,`image`,`colorKey`) VALUES ('".$d['name']."','".$d['notes']."','".$d['image']."','".$d['colorKey']."')");
+				
+		}
+				
+		if($s){
+			$result['success'] = "Import complete";
+			unlink($target_path);
+		}else{
+			$result['error'] = "There was an error importing your JSON file ".mysqli_error($conn);
+			echo json_encode($result);
+			return;
+		}
+			
+	} else {
+		$result['error'] = "There was an error processing json file $target_path, please try again!";
+		//echo json_encode($result);
+
+	}
+	echo json_encode($result);
+	return;
+
+}
+
+
 //EXPORT INGREDIENT CATEGORIES
 if($_GET['action'] == 'exportIngCat'){
 	if(empty(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM ingCategory")))){