+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v10.3
 --------------------------
+- Added JSON file import for ingredient categories
 - Further improvements in SDS generation
 - Ingredients safety info page updates
 - For full details please refer to the CHANGELOG